refact: separating code for keyboard-shortcuts feature.
