+++ packages/wxt/src/types.ts
@@ -535,11 +535,6 @@ export interface WxtDevServer
   restartBrowser: () => void;
 }
 
-export interface KeyboardShortcutWatcher {
-  start(): void;
-  stop(): void;
-}
-
 export interface ReloadContentScriptPayload {
   registration?: BaseContentScriptEntrypointOptions['registration'];
   contentScript: Omit<Scripting.RegisteredContentScript, 'id'>;
