--- sh.exe.stackdump
@@ -1,29 +1,0 @@
-Stack trace:
-Frame         Function      Args
-0007FFFFC060  00021005FE8E (000210285F68, 00021026AB6E, 0007FFFFC060, 0007FFFFAF60) msys-2.0.dll+0x1FE8E
-0007FFFFC060  0002100467F9 (000000000000, 000000000000, 000000000000, 0007FFFFC338) msys-2.0.dll+0x67F9
-0007FFFFC060  000210046832 (000210286019, 0007FFFFBF18, 0007FFFFC060, 000000000000) msys-2.0.dll+0x6832
-0007FFFFC060  000210068CF6 (000000000000, 000000000000, 000000000000, 000000000000) msys-2.0.dll+0x28CF6
-0007FFFFC060  000210068E24 (0007FFFFC070, 000000000000, 000000000000, 000000000000) msys-2.0.dll+0x28E24
-0007FFFFC340  00021006A225 (0007FFFFC070, 000000000000, 000000000000, 000000000000) msys-2.0.dll+0x2A225
-End of stack trace
-Loaded modules:
-000100400000 sh.exe
-7FFFA7450000 ntdll.dll
-7FFFA6780000 KERNEL32.DLL
-7FFFA4BA0000 KERNELBASE.dll
-7FFFA5EC0000 USER32.dll
-7FFFA4750000 win32u.dll
-000210040000 msys-2.0.dll
-7FFFA5D10000 GDI32.dll
-7FFFA48C0000 gdi32full.dll
-7FFFA4F60000 msvcp_win.dll
-7FFFA49E0000 ucrtbase.dll
-7FFFA6F40000 advapi32.dll
-7FFFA5E10000 msvcrt.dll
-7FFFA7070000 sechost.dll
-7FFFA4B70000 bcrypt.dll
-7FFFA5370000 RPCRT4.dll
-7FFFA3EF0000 CRYPTBASE.DLL
-7FFFA4780000 bcryptPrimitives.dll
-7FFFA51D0000 IMM32.DLL
